feat: Add initial Rubocop configuration for test directory
This commit introduces a `.rubocop.yml` file to the `unity/test`
directory. This file configures Rubocop, a Ruby static code analyzer,
to enforce specific coding styles and conventions within the test suite.

The configuration disables several rules that conflict with the
existing ThrowTheSwitch coding style or are considered too strict for
the current stage of development.  These include:

- Style/SpecialGlobalVars:  Allow perl names for globals.
- Style/FormatString, Style/GlobalVars, Style/FrozenStringLiteralComment, Style/NumericPredicate, Style/MultilineBlockChain, Style/Alias, Style/EvalWithLocation, Style/MixinUsage, Style/OptionalBooleanParameter: Disable specific style rules.
- Layout/HashAlignment, Layout/LineLength: Disable layout rules.
- Security/Eval, Security/YAMLLoad: Disable security checks.
- Style/Documentation, Style/DocumentationMethod: Disable documentation rules.
- Metrics/*: Disable all metric checks.

This initial configuration aims to provide a baseline for future
improvements in code quality and consistency.